+++ lib/config.js
@@ -1,4 +1,4 @@
-export default {
+export const config = {
     overpassBase: '//overpass-cfn-production.tilestream.net/api/interpreter',
     osmBase: '//www.openstreetmap.org/api/0.6/',
     mapboxAccessToken: 'pk.eyJ1IjoicmFzYWd5IiwiYSI6ImNpejVrMjc4eTAwNGczM2thNWozYnJ1OHkifQ.yFRr3Sd39TJiwEguQpIkWQ'
