+++ lib/getChangeset.js
@@ -1,7 +1,7 @@
 import xhr from 'xhr';
 import adiffParser from 'osm-adiff-parser-saxjs';
 import jsonParser from 'real-changesets-parser';
-import { query } from './osm';
+import { query } from './query';
 
 var S3_URL = '//s3.amazonaws.com/mapbox/real-changesets/production/';
