+++ lib/index.js
@@ -1,0 +1,5 @@
+export { config } from './config';
+export { getChangeset } from './getChangeset';
+export { query } from './query';
+export { propsDiff } from './propsDiff';
+export { render } from './render';
