+++ lib/render.js
@@ -3,12 +3,13 @@ import moment from 'moment';
 import { EventEmitter as events } from 'events';
 import turfBboxPolygon from '@turf/bbox-polygon';
 import helpers from '@turf/helpers';
+
 import { getChangeset } from './getChangeset';
 import { propsDiff } from './propsDiff';
 
 const featureCollection = helpers.featureCollection;
 
-import config from './config';
+import { config } from './config';
 
 var cmap, map;
