+++ package.json
@@ -2,9 +2,9 @@
   "name": "cmap",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "dist/bundle.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build website": "rollup -c ./rollup.website.config.js"
   },
   "author": "",
   "license": "ISC",
@@ -25,7 +25,6 @@
     "babel-plugin-external-helpers": "^6.22.0",
     "babel-preset-es2015-rollup": "^3.0.0",
     "babel-preset-react": "^6.24.1",
-    "eslint-config-mourner": "^2.0.1",
     "eslint-plugin-react": "^7.0.0",
     "rollup": "^0.41.6",
     "rollup-plugin-babel": "^2.7.1",
